feet : 전자결재 승인 시, 보유휴가 차감되도록 완료.

+++ HRM_System/src/main/java/com/min/edu/dto/LeaveDto.java
@@ -19,6 +19,6 @@ public class LeaveDto {
 	private String leave_start;
 	private String leave_end;
 	private String type;
-	private int leave_days;
+	private double leave_days;
 
 }
